Refactor: Move movement logic from Unit to Mover
