+++ src/main/java/website/frontrow/util/CollisionHandler.java
@@ -6,6 +6,7 @@ import website.frontrow.board.Bubble;
 import website.frontrow.board.Enemy;
 import website.frontrow.board.Player;
 import website.frontrow.board.Unit;
+import website.frontrow.board.Mover;
 import website.frontrow.level.Level;
 
 /**
@@ -28,9 +29,9 @@ public class CollisionHandler
     }
     
     /**
-     * Create an ArrayList with box values for a unit.
+     * Create an ArrayList with box values for a mover.
      * @param location
-     * Input the location point of a Unit
+     * Input the location point of a mover
      * @return
      * Return an ArrayList of all box values.
      */
@@ -45,20 +46,19 @@ public class CollisionHandler
     }
 
     /**
-     * Check if a unit collides with another unit. Call collisionApplier if it does.
-     * The idea is that we check the units projected location against the boxed
-     * locations of all other units in a loop.
-     * If any of the box corners of the moving unit falls within
-     * the box of another unit there is a collision.
-     * There is also a check to filter whether the unit is colliding with itself.
-     * @param loc
-     * The location of the current unit
+     * Check if a mover collides with another mover. Call collisionApplier if it does.
+     * The idea is that we check the movers projected location against the boxed
+     * locations of all other movers in a loop.
+     * If any of the box corners of the moving mover falls within
+     * the box of another mover there is a collision.
+     * There is also a check to filter whether the mover is colliding with itself.
+	 * @param loc
+     * The location of the current mover
      * @param mov
-     * The move the current unit wants to make
-     * @param unit
-     * The unit that is moving
-     */
-    public void checkUnitCollision(Point loc, Point mov, Unit unit)
+     * The move the current mover wants to make
+	 * @param mover The mover to check with.
+	 */
+    public void checkMoverCollision(Point loc, Point mov, Mover mover)
     {
     	loc.add(mov);
     	ArrayList<Point> aabbPlayer = buildBox(loc);
@@ -75,10 +75,10 @@ public class CollisionHandler
 	    		if      (currentX >= otherX
                         && currentY >= otherY
                         && currentX < (otherX + 1) && currentY < (otherY + 1)
-                        && other.get(k) != unit)
+                        && other.get(k) != mover)
 	    		{
 
-					applyCollision(unit, other.get(k));
+					applyCollision(mover, other.get(k));
 	    		}
 	    	}
 		}   	
@@ -86,11 +86,11 @@ public class CollisionHandler
 
 	/**
 	 * Call the method that handles collisions.
-	 * @param collider The unit that initiates the collision.
-	 * @param colidee The unit that is being collided with.
+	 * @param collider The mover that initiates the collision.
+	 * @param colidee The mover that is being collided with.
 	 */
 	// TODO: Swap this out for something fancy that uses reflection. But this works fine for now.
-	public void applyCollision(Unit collider, Unit colidee)
+	public void applyCollision(Mover collider, Mover colidee)
 	{
 		if(collider instanceof Player)
 		{
@@ -112,21 +112,21 @@ public class CollisionHandler
 	}
 
 	/**
-	 * Called when a player collides with another unit.
+	 * Called when a player collides with another mover.
 	 * @param player Player which is currently colliding.
-	 * @param other The unit that was collided with.
+	 * @param other The mover that was collided with.
 	 */
-	public void playerCollision(Player player, Unit other)
+	public void playerCollision(Player player, Mover other)
 	{
 		// TODO: Make the player lose a life or something else.
 	}
 
 	/**
-	 * Called when a bubble collides with another unit.
+	 * Called when a bubble collides with another mover.
 	 * @param bubble Bubble which is currently colliding.
-	 * @param other The unit that was collided with.
+	 * @param other The mover that was collided with.
 	 */
-	public void bubbleCollision(Bubble bubble, Unit other)
+	public void bubbleCollision(Bubble bubble, Mover other)
 	{
 		if(other instanceof Enemy)
 		{
@@ -138,17 +138,17 @@ public class CollisionHandler
 	}
 
     /**
-     * Checks if a units projected movement makes it collide with a cell that is non-empty.
+     * Checks if a movers projected movement makes it collide with a cell that is non-empty.
      * @param loc
-     * The location of the current unit
+     * The location of the current mover
      * @param mov
-     * The move the current unit wants to make
-     * @param unit
-     * The unit that is moving
+     * The move the current mover wants to make
+     * @param mover
+     * The mover that is moving
      * @return
      * Whether there has been a collision or not
      */
-    public boolean checkCellCollision(Point loc, Point mov, Unit unit)
+    public boolean checkCellCollision(Point loc, Point mov, Mover mover)
     {
     	ArrayList<Point> aabb = buildBox(loc);
     
@@ -161,7 +161,7 @@ public class CollisionHandler
             {
     			return true;
     		}			
-    		if(level.getCells().get(x, y).collides(unit))
+    		if(level.getCells().get(x, y).collides(mover))
     		{ 			
     			return true;
     		}
